+++ apps/cli/src/adapters/commander/commands/info.ts
@@ -1,13 +1,18 @@
 import { Command } from "commander";
 
-import { Dependencies } from "../../../domain/dependencies.js";
-import { getInfo, printInfo } from "../../../domain/info.js";
+import type { Dependencies } from "../../../domain/dependencies.js";
+
+import { getInfo } from "../../../domain/info.js";
+import { GlobalOptions } from "../global-options.js";
+import { createCommandPresenter } from "../presenters/index.js";
 
 export const makeInfoCommand = (dependencies: Dependencies): Command =>
   new Command()
     .name("info")
     .description("Display information about the project")
-    .action(async () => {
+    .action(async function () {
+      const { output } = this.optsWithGlobals<GlobalOptions>();
+      const presenter = createCommandPresenter(output);
       const result = await getInfo(dependencies)();
-      printInfo(result);
+      presenter.reportResult(result);
     });